+++ CHANGELOG.md
@@ -59,6 +59,7 @@ In this release we had a little focus on:
 - Forbids to use ``open()`` outside of ``with``
 - Forbids to use ``type()`` for compares
 - Forbids to have consecutive expressions with too deep access level
+- Forbids to have too many public instance attributes
 
 ### Bugfixes